Refactor custom widgets module
