+++ simple-us/utils/widgets/checkbox.py
@@ -1,31 +1,6 @@
 from typing import Dict
 
-from ipymaterialui import Icon
-from ipymaterialui import IconButton
-from ipymaterialui import Html
-from ipywidgets import jslink
-
-
-class CustomText(Html):
-    """ Custom text class to make sure the default text size is consistent.
-
-    The default text size on Jupyter Lab and Jupyter Notebook is not consistent.
-    """
-
-    def __init__(self, text, tag="div", style_: Dict = None, **kwargs):
-        super(Html, self).__init__()
-
-        if style_ is None:
-            style_ = {
-            }
-        if "padding" not in style_.keys():
-            style_["padding"] = "0px 0px 0px 0px"
-        if "font-size" not in style_.keys():
-            style_["font-size"] = 15
-
-        self.tag = tag
-        self.children = text
-        self.style_ = style_
+from ipymaterialui import IconButton, Icon
 
 
 class CustomCheckbox(IconButton):
@@ -74,5 +49,3 @@ class CustomCheckbox(IconButton):
             self.children = self._checked_icon
         elif not self._checked:
             self.children = self._unchecked_icon
-
-
